Enable Jest coverage in SonarCloud (#1)
* Remove sonar-project.properties and use pom.xml

The SonarCloud GitHub Action complains about it being a maven project. Properties
from sonar-project.properties have been moved to the pom.xml.

* Revert "Remove sonar-project.properties and use pom.xml"

This reverts commit 5d75ce1a6a9582c4a778a26bf768b4fa5c4e9799.

* Revert back to using the GitHub Action

SonarCloud's GitHub Action requires a CI-based workflow, however it will
not work if the repo is a maven project. A step is added to delete the
pom.xml in order to bypass this check. Properties from
sonar-project.properties has been added to the pom.xml.

* Add sonar.java.binaries to project properties

* mvn clean package in workflow